backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
httpx==0.28.1
python-dotenv==1.0.1
slowapi==0.1.9
nbtlib==2.0.4
redis==5.0.8
